applicatie/requirements.txt
Click==8.1.2
cycler==0.11.0
Flask==2.1.1
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.1
Pillow==9.1.0
pandas==1.4.2
scikit-learn==1.0.2
joblib=1.1.0
